Install behavior
Added a break, I mean, you go out of your way to progress the download, I figure the install progress bar is two to much, some installers go into temp first and it all seemed ridiculous, so I have it pausing and then going on just install right -of-way.  I decided to put a break indefinite on starting the install or uninstall.