MaxLand for Visual Basic Applications(R)
I have integrated the VBA IDE into the recent VBScript changes and object exposures, I don't know how practical it is but maybe someone else knows better it's capabilities and could find use of it.  Certainly not necessary to VBScript the engine now and create and access it all through objects and collections, the Word Document help for the language reference has been thoroughly updated.