I stand corrected.
CoPilot didn't hint on the center/normal/lengths theory, just the small code part, but it just generated code that doesn't work.  However it's VB6 version of the matter was sufficient to converted to C++ and when the Collision() function is redeveloped with it, I will be curious to see the benchmark difference.  Should be interesting.